requirements.txt
langchain>=0.2.13,<0.3.0
PyPDF2
python-dotenv
streamlit>=1.18.1,<1.33.0
faiss-cpu
streamlit-extras
langchain-community
tiktoken
reportlab
pandas
PyMuPDF
selenium
textblob
sumy
googlesearch-python
openai
bs4
Pillow
plotly
newspaper3k
python-docx
fitz
nltk
feedparser
requests
nest_asyncio
nemoguardrails
